Use new gzip.Writer cache in rest/multipart.go
Had to move the cache code down to the base package.
